+++ tests/files-test.js
@@ -1,12 +1,11 @@
 import {
-    textFileContent,
     writeTextInFile,
     concatenateFiles,
     copyFile,
     copyDirectory,
     deleteFile
 } from "../files.js";
-
+import fsPromises from "fs/promises";
 
 const initialString = `abc - éùô - xwz`;
 
@@ -22,8 +21,21 @@ const pathDeep = `${a}/${b}/c.txt`;
 //         await deleteFile(pathDeep);
 //     });
 // });
+// writeTextInFile(pathDeep, initialString)
+fsPromises.writeFile(pathDeep, initialString+"uuuuu");
+// textFileContent("tests/folder/text.txt",).then(async function (result) {
+//     // console.log(`success`, result === initialString);
+//     console.log(result);
 
-copyDirectory(`tests`, `myCopy`)
-
+//     // await deleteFile(pathDeep);
+// });
+// fsPromises.readFile("tests/folder/text.txt","utf-8").then(async function (result) {
+//             // console.log(`success`, result === initialString);
+//             console.log(result);
+    
+//             // await deleteFile(pathDeep);
+//         });
+// copyDirectory(`tests`, `myCopy`)
+// copyFile("tests/folder/text.txt", "tests/xxx/copy.txt").then(console.log)
 
 // copyFile(`tests/files-test.js`, `copyFile/files-test.js`)